# laravel11/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SyncBridge API",
        "description": "Production-grade API documentation for SyncBridge integration endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Primary API server"
        }
    ],
    "paths": {
        "/api/legacy": {
            "get": {
                "tags": [
                    "Legacy"
                ],
                "summary": "Legacy endpoint (GET)",
                "description": "Legacy compatibility endpoint. Uses action query parameter to route behavior.",
                "operationId": "legacyHandleGet",
                "parameters": [
                    {
                        "name": "action",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "status",
                                "products",
                                "orders",
                                "settings",
                                "settings_save",
                                "sync"
                            ]
                        }
                    },
                    {
                        "name": "command",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "products"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Legacy action succeeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unsupported legacy action",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Legacy sync failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Legacy"
                ],
                "summary": "Legacy endpoint (POST)",
                "description": "Legacy compatibility endpoint. Accepts body payload for action dispatching.",
                "operationId": "legacyHandlePost",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "enum": [
                                            "status",
                                            "products",
                                            "orders",
                                            "settings",
                                            "settings_save",
                                            "sync"
                                        ]
                                    },
                                    "command": {
                                        "type": "string",
                                        "example": "products"
                                    },
                                    "settings": {
                                        "type": "object",
                                        "additionalProperties": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Legacy action succeeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unsupported legacy action",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Legacy sync failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiError": {
                "required": [
                    "ok",
                    "error",
                    "errors"
                ],
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Invalid payload"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "OpenApiSpec",
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste SYNCBRIDGE_API_TOKEN as Bearer token.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            },
            "apiTokenHeader": {
                "type": "apiKey",
                "description": "Alternative auth header for sync-protected endpoints.",
                "name": "X-Api-Token",
                "in": "header"
            }
        }
    },
...
}